# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Esta é uma API simples usando Gin e Swagger.",
        "title": "Golang Documentação API 2.5",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:4000",
    "paths": {
        "/auth/cadastro-externo": {
            "post": {
                "description": "Autentica um usuário externo e retorna um token JWT com os dados do usuário.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Autenticação"
                ],
                "summary": "Cadastro de Usuário Externo",
                "parameters": [
                    {
                        "description": "Dados de Cadastro",
                        "name": "cadastro",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/userDto.UserSignin"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Erro ao processar a requisição",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "description": "Autentica um aluno e retorna um token JWT com os dados do usuário.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Autenticação"
                ],
                "summary": "Autenticação de Usuário Aluno",
                "parameters": [
                    {
                        "description": "Dados de Login",
                        "name": "login",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/userDto.UserLogin"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Erro ao processar a requisição",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/auth/login-externo": {
            "post": {
                "description": "Autentica um usuário externo e retorna um token JWT com os dados do usuário.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Autenticação"
                ],
                "summary": "Autenticação de Usuário Externo",
                "parameters": [
                    {
                        "description": "Dados de Login",
                        "name": "login",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/userDto.UserLogin"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Erro ao processar a requisição",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/postagens/listar": {
            "get": {
                "description": "Gera uma lista paginada com as postagens do blog",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Postagens"
                ],
                "summary": "Pegar listagem de postagens",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Pagina",
                        "name": "pagina",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Limite por página",
                        "name": "limite",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Pesquisa",
                        "name": "pesquisa",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listagem a mostra:",
                        "schema": {
                            "$ref": "#/definitions/postagensDto.ListagemPostagens"
                        }
                    },
                    "400": {
                        "description": "Erro ao processar a requisição",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/postagens/postar": {
            "post": {
                "description": "Cria uma nova postagem no blog",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Postagens"
                ],
                "summary": "Criar uma nova postagem",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Dados da nova postagem",
                        "name": "postagem",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/postagensDto.NovaPostagem"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Postagem criada com sucesso",
                        "schema": {
                            "$ref": "#/definitions/postagensDto.PostagemDataComplete"
                        }
                    },
                    "400": {
                        "description": "Erro ao processar a requisição",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/postagens/{postagemId}/comentar": {
            "post": {
                "description": "Adiciona um comentário a uma postagem existente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Comentários"
                ],
                "summary": "Adicionar um comentário a uma postagem",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID da postagem",
                        "name": "postagemId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados do comentário",
                        "name": "comentario",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/postagensDto.ComentarioData"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Comentário inserido com sucesso",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Erro ao processar a requisição",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/postagens/{postagemId}/reagir": {
            "post": {
                "description": "Adiciona ou atualiza uma reação (LIKE, FOGUINHO, AMEI, ODIEI) para uma postagem específica",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reações"
                ],
                "summary": "Adicionar ou atualizar uma reação a uma postagem",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID da postagem",
                        "name": "postagemId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Dados da reação",
                        "name": "reacao",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/postagensDocDto.ReacaoData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reação atualizada com sucesso",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "201": {
                        "description": "Reação inserida com sucesso",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Erro ao processar a requisição",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/postagens/{titulo}": {
            "get": {
                "description": "Pega uma postagem por titulo",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Postagens"
                ],
                "summary": "Pegar postagem por titulo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Título da postagem",
                        "name": "titulo",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Tipo de postagem: (alunoPost,ufcaPost)",
                        "name": "tipo",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Postagem a mostra",
                        "schema": {
                            "$ref": "#/definitions/postagensDto.PostagemDataComplete"
                        }
                    },
                    "400": {
                        "description": "Erro ao processar a requisição",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "postagensDocDto.ReacaoData": {
            "type": "object",
            "properties": {
                "tipo": {
                    "type": "string",
                    "example": "LIKE"
                }
            }
        },
        "postagensDto.ComentarioData": {
            "type": "object",
            "properties": {
                "conteudo": {
                    "type": "string"
                },
                "criadoEm": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "parenteId": {
                    "type": "integer"
                },
                "postagemId": {
                    "description": "Validado na url",
                    "type": "integer"
                },
                "usuarioId": {
                    "type": "integer"
                }
            }
        },
        "postagensDto.ComentarioDataComplete": {
            "type": "object",
            "properties": {
                "autor": {
                    "$ref": "#/definitions/postagensDto.UserPostagem"
                },
                "conteudo": {
                    "type": "string"
                },
                "criadoEm": {
                    "type": "string"
                },
                "filhos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/postagensDto.ComentarioDataComplete"
                    }
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "postagensDto.ListagemPostagens": {
            "type": "object",
            "properties": {
                "TotalPaginas": {
                    "type": "integer"
                },
                "limite": {
                    "type": "integer"
                },
                "pagina": {
                    "type": "integer"
                },
                "pesquisa": {
                    "type": "string"
                },
                "postagem": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/postagensDto.PostagemDataLista"
                    }
                }
            }
        },
        "postagensDto.NovaPostagem": {
            "type": "object",
            "properties": {
                "conteudo": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "tipo": {
                    "type": "string"
                },
                "titulo": {
                    "type": "string"
                },
                "usuarioId": {
                    "type": "integer"
                }
            }
        },
        "postagensDto.PostagemDataComplete": {
            "type": "object",
            "properties": {
                "autor": {
                    "$ref": "#/definitions/postagensDto.UserPostagem"
                },
                "comentarios": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/postagensDto.ComentarioDataComplete"
                    }
                },
                "conteudo": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "reacoes": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "tipo": {
                    "type": "string"
                },
                "titulo": {
                    "type": "string"
                }
            }
        },
        "postagensDto.PostagemDataLista": {
            "type": "object",
            "properties": {
                "autor": {
                    "$ref": "#/definitions/postagensDto.UserPostagem"
                },
                "comentarios": {
                    "type": "integer"
                },
                "conteudo": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "reacoes": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "tipo": {
                    "type": "string"
                },
                "titulo": {
                    "type": "string"
                }
            }
        },
        "postagensDto.UserPostagem": {
            "type": "object",
            "properties": {
                "contexto": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "login": {
                    "type": "string"
                },
                "nome": {
                    "type": "string"
                }
            }
        },
        "userDto.UserLogin": {
            "type": "object",
            "properties": {
                "login": {
                    "type": "string"
                },
                "senha": {
                    "type": "string"
                }
            }
        },
        "userDto.UserSignin": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "login": {
                    "type": "string"
                },
                "nome": {
                    "type": "string"
                },
                "senha": {
                    "type": "string"
                }
            }
        }
    }
}